fix(codegen): define out-of-range casts and min literals
Float-to-integer `as` lowered to fptosi/fptoui, which yield poison for
any value the target cannot represent. The program's output stopped
being a function of its source: 1e300 as i32 printed -2147483648 at -O0,
printed nothing at all at -O3, and the constant folder computed a third
answer again. Lower through llvm.fptosi.sat / llvm.fptoui.sat instead,
and compute the same total function in FoldedConst::cast_to so a folded
cast and a run-time one agree.

The most negative value of a signed type had no literal spelling. A
negation is an operator over a literal, so the checker range-checked the
magnitude, which is one past the type's maximum; and the lexer carried
every literal in an i64, which excluded i64::MIN and the upper half of
u64 before a type was known. Integer tokens now carry an unsigned
magnitude, Literal::Integer carries an i128, and a negation over a
literal is checked against the negated value.

Four parser sign guards go with it as unreachable: no integer regex
admits a sign, so they could never fire.

+++ compiler/hir-lowering/src/iteration.rs
@@ -392,7 +392,7 @@ fn variable(name: &str, ty: HirType, span: Span) -> HirExpr {
 
 fn int_literal(value: u64, ty: HirType, span: Span) -> HirExpr {
     HirExpr::new(
-        HirExprKind::Literal(Literal::Integer(value as i64, None)),
+        HirExprKind::Literal(Literal::Integer(value as i128, None)),
         ty,
         span,
     )
